feat: Update list tags structure
