+++ src/script/components/panel/PanelActions/PanelActions.tsx
@@ -39,6 +39,7 @@ const PanelActions: React.FC<PanelActionsProps> = ({items}) => {
     <ul css={listCSS}>
       {items.map(({click, identifier, icon, label}) => (
         <li key={identifier}>
+          1st one
           <button className="panel__action-item" onClick={click} data-uie-name={identifier} type="button">
             <span className="panel__action-item__icon">
               <Icon name={icon} />
