+++ src/style/panel/panel.less
@@ -158,6 +158,7 @@
     }
 
     &--head {
+      display: block;
       font-weight: @font-weight-bold;
       text-transform: uppercase;
     }
